Cancelling of (running) background job + various other changes
Changes:
- Actions can now be scheduled on executing background jobs
- Properly setting of utc kind for objects returned from storage
- Fixed bug in MySqlDequeuedJob where ExpectedTimeoutUtc was not updated after heartbeat
- Re-added IStorageConnection parameter to IJobQueue.EnqueueAsync
- BackgroundJobExecutionContext now logs to daemon
- BackgroundJobExecutionContext now handles pending actions
- Added option to cancel executing job on IBackgroundJobExecutionContext
- WorkerSwarmHost now requeues cancelled jobs
- Added Async to method names related to IActivator
- Removed LockHeartbeatSafetyOffset from background job swarm options in favor of the offset configured for the HiveMind environment
- Changed middlware priority from ushort to byte
- Renamed background job QueryAsync to SearchAsync
- Renamed background job QueryCountAsync to CountAsync
- More improvements to BackgroundJobAwaitingProcessTrigger
- Added CancelBackgroundJobAction
- Added ActionInfo

+++ src/Sels.Hivemind/Sels.HiveMind.Core/Contracts/IMiddlewareInfo.cs
@@ -20,6 +20,6 @@ namespace Sels.HiveMind
         /// <summary>
         /// Optional priority of the middleware to determine the execution order. Lower priority means executed first, null means executed last in the chain.
         /// </summary>
-        public uint? Priority { get; }
+        public byte? Priority { get; }
     }
 }
